requirements.txt
altgraph==0.17.4
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
argcomplete==3.6.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
fastapi==0.120.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
lmstudio==1.5.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mpmath==1.3.0
msgspec==0.19.0
networkx==3.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pefile==2023.2.7
pillow==11.0.0
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.3
pytesseract==0.3.13
python-barcode==0.16.1
python-dateutil==2.9.0.post0
python-escpos==3.1
python-multipart==0.0.20
pytz==2025.2
pyusb==1.3.1
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qrcode==8.2
requests==2.32.4
scipy==1.16.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
sympy==1.13.3
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.178
ultralytics-thop==2.0.15
urllib3==1.26.13
uvicorn==0.38.0
wsproto==1.2.0
